+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Version latest (0.12.0rc?, 2019-09-24?)
 	- pass `is_outdated` in click context for error verbosity about upgrade
 - feat: use click's Command to display footer after invoke
 - feat: unit tests for the above changes
+- feat: use click's UsageError for the deprecated options
 
 
 Version 0.12.{0,1} (2019-10-25)